pyproject.toml
[project]
name = "tak-feeder-nysse"
version = "0.1.0"
description = ""
authors = [
    {name = "Gemini CLI"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "pytak (>=7.2.1,<8.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "click (>=8.3.1,<9.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "aiohttp (>=3.13.3,<4.0.0)"
]

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "types-requests (>=2.32.4.20260107,<3.0.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pylint.messages_control]
disable = []

[tool.pylint.format]
max-line-length = 88
